pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["Andrey Tumerkin <tumerrkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
load-dotenv = "^0.1.0"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pillow = "^11.1.0"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
coverage = "^7.6.10"
drf-yasg = "^1.21.8"
ipython = "^8.31.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"